+++ src/components/Tabbar/Tabbar.jsx
@@ -1,0 +1,12 @@
+import TabbarContainer from './Tabbar.styled';
+
+const Tabbar = () => {
+	return (
+		<TabbarContainer>
+			<button>SAVED</button>
+			<button>GOING</button>
+		</TabbarContainer>
+	);
+};
+
+export default Tabbar;
