+++ src/components/Tabbar/Tabbar.styled.jsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const TabbarContainer = styled.section`
+	display: flex;
+	width: 100vw;
+	height: auto;
+
+	button {
+		width: 50%;
+	}
+`;
+
+export default TabbarContainer;
